ARM: KVM: Add hypervisor inititalization
Sets up the required registers to run code in HYP-mode from the kernel.
No major controversies, but we should consider how to deal with SMP
support for hypervisor stack page.

Works by setting the HVBAR the kernel can execute code in Hyp-mode with
the MMU disabled which initializes other registers and enables the MMU
for Hyp-mode.

Unfortunately, this commit doesn't yet work as the configured
translation is somehow faulty.