ARM: KVM: Move interrupt handling outside of interrupts.S
While we have to take host interrupts in HYP mode while running a
guest, calling __irq_svc ourself is quite controvertial.

And actually, we can let the kernel do its magic without getting in
the way. Let's just return to SVC and enable interrupts. This should
be enough for the CPU to notice that an interrupt is pending, and
handle the interrupt the normal way.

Maybe we should consider renaming interrupts.S, after all... ;-)

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>