ARM: KVM: limit the use of .arch_extension sec/virt
Having the .arch_extension sec/virt directives in unified.h is not
very nice, as it enables the extension even in files having nothing
to do with secure mode or virtualization.

Instead, move the selection of the sec/virt extensions into the KVM
Makefile. This is in line with the current practice in the kernel,
with the additional benefits of allowing compilation with older
toolchains and touching one less core file.

Spell-checked-by: Peter Maydell <peter.maydell@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>