ARM: KVM: Change emulation table to a register table.
The idea is that each register we care about in the guest has a single
entry in the table.  This table can then drive our set/get API for
userspace as well as emulation and reset.

This means we fold the read and write functionality into the accessor
functions.  We rename struct coproc_emulate to struct coproc_reg
to reflect the more generic role, and the coproc_emulate[] table to
the more precise 'cp15_regs[]'.

Conflicts:

	arch/arm/kvm/emulate.c

Signed-off-by: Rusty Russell <rusty.russell@linaro.org>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>