ARM: KVM: Switch HVBAR setting to be HVC-based
Instead of relying on SMC calls to setup HVBAR, introduce a new
HVC-based API. This of course requires a hypervisor stub to be
installed already. At the moment, this is implemented as part of the
boot wrapper, but will be moved to the kernel itself when it actually
understands being booted in HYP mode.

The kernel now uses "hvc #0xff", with r0 containing the pointer to
the new vectors.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>