KVM ARM: Use KVM_ARM_VCPU_INIT to set target model and features.
This uses a fixed feature bitmap for simplicity: 224 features should
be enough for anyone!

QEMU now needs to explicitly state what CPU a guest vcpu has, and what
optional features (currently, none are defined, but this could be used
to suppress VFP, debug, etc).

Once we know the target (which currently must be 0xCOF), we do the
vcpu reset; trying to run without that will -ENOEXEC.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>