ARM: KVM: vfp: Use a per-CPU struct for host VFP state
Using a vfp_hard_struct per vcpu for the host is a bit overkill
(odds are we have more vcpus than physical CPUs). It then makes
sense to have a global set of vfp_hard_structs for the host.

Define a per-cpu variable holding the host CPU states, and let
each vcpu point to the right state as they're about to run.

Cc: Antonios Motakis <a.motakis@virtualopensystems.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>