KVM: ARM: Atomically read guest instruction,
Ben Herrenschmidt pointed out a nasty race:
(1) Guest complicated mmio instruction traps.
(2) The hardware doesn't tell us enough, so we need to read the actual
    instruction which was being exectuted.
(3) KVM maps the instruction virtual address to a physical address.
(4) The guest (SMP) swaps out that page, and fills it with something else.
(5) We read the physical address, but now that's the wrong thing.

This is so rare that we can stop all the vcpus from running when it
happens (if no VCPUS are running, this just means getting and
releasing a spinlock).

Signed-off-by: Rusty Russell <rusty.russell@linaro.org>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>