ARM: KVM: simplify stack layout in HYP mode
On world switch, we save the pointer to the current vcpu on the top
of the stack. This tends to complicate the exit path a bit as we then
need to dig into the stack to restore it (we've already pushed a number
of registers on top of it at that time), not to mention error prone.

A obvious way to simplify this is to use the HTPIDR register, which is an
architected method to store per-CPU data in HYP mode, and this is
exactly what this patch does.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>