KVM: ARM: Inject data/prefetch aborts in guest
Inject these aborts if the guest tries to do something really weird.
It's nicer to tell the guest that there was an abort than just
forcefully kill the whole thing.

For now we always tell the guest it was a debug generated abort, but we
can make this scheme more flexible in the future.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>