Revert "ARM: KVM: move usr regs to struct pt_regs"
This partially reverts commit f7c6882646ebd12ee0daab1055cd38c23f47c610.

The kvm_regs struct is part of the user space ABI to provide an index
for which registers are get/set using the corresponding ioctl's and
should really be explicitly typed in line with the remaining KVM code.

If we want to move away from that we should provide fixed indices for
the registers, but I've never been a fan of hiding the pc and cpsr under
the usr_regs field in the first place.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>