KVM: ARM: Check for overlaps of mapped io addresses
When calling stage2_set_pte from kvm_phys_addr_ioremap we pass an
argument to say that this is an IO mapping, and that we expect the
adress range to be free, otherwise return an error.

This should catch errors earlier when user space supplies guest physical
addresses that overlap.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>